Separating emulator into components
